feat:merge docstubs if both source and docstubs have them
