+++ pyproject.toml
@@ -50,7 +50,7 @@ python-minifier = "^2.7.0"
 requests = "^2.28.0"
 mypy = "1.7.0"
 black = ">=22.10,<23.0"
-libcst = ">=0.4.9,<1.2.0"
+libcst = "^1.1.0"
 autoflake = ">=1.7,<3.0"
 cachetools = "^5.3.0"
 click = "^8.1.3"
