+++ src/stubber/cst_transformer.py
@@ -43,24 +43,15 @@ class StubTypingCollector(cst.CSTVisitor):
         # store the annotations
         self.annotations: Dict[
             Tuple[str, ...],  # key: tuple of canonical class/function name
-            TypeInfo,  # value: TypeInfo
+            Union[TypeInfo, str],  
         ] = {}
 
     # ------------------------------------------------------------
     def visit_Module(self, node: cst.Module) -> bool:
         """Store the module docstring"""
-        if node.get_docstring():
-            ## TODO: try / catch
-            assert isinstance(node.body[0], cst.SimpleStatementLine)
-            ti = TypeInfo(
-                name="module",
-                params=None,
-                returns=None,
-                docstr_node=node.body[0],
-                decorators=(),
-                def_type="module",
-            )
-            self.annotations[MODULE_KEY] = ti
+        docstr = node.get_docstring()
+        if docstr:
+            self.annotations[MODULE_KEY] = docstr
         return True
 
     # ------------------------------------------------------------
