+++ tests/codemods/codemod_test_cases/class_init_kwargs/doc_stub.py
@@ -1,5 +1,5 @@
 """
-__init__ arguments from docstubs
+__init__ arguments
 """
