+++ tests/codemods/codemod_test_cases/class_init_kwargs/expected.py
@@ -1,5 +1,5 @@
 """
-__init__ arguments from docstubs
+__init__ arguments
 """
