+++ tests/codemods/codemod_test_cases/import_all_add/before.py
@@ -1,5 +1,5 @@
 """
-add import
+add all import styles
 """
