+++ tests/codemods/codemod_test_cases/mod_docstring_update/before.py
@@ -1,4 +1,4 @@
-"before mod docstring"
+"before module docstring"
 
 def foo():
     ''
