+++ tests/codemods/codemod_test_cases/mod_docstring_update/expected.py
@@ -1,5 +1,8 @@
 """
-module docstring
+docstub module docstring
+
+---
+before module docstring
 """
 
 def foo():
