+++ tests/codemods/codemod_test_cases/mod_docstring_update_same_1/before.py
@@ -1,6 +1,5 @@
-"""
-module docstring
-"""
+"The same module docstring."
+
 def foo():
     ''
     ...
