+++ tests/codemods/codemod_test_cases/mod_docstring_update_same_1/doc_stub.py
@@ -1,5 +1,5 @@
 """
-docstub module docstring
+The same module docstring.
 """
