+++ tests/codemods/codemod_test_cases/mod_docstring_update_same_3/doc_stub.py
@@ -1,5 +1,5 @@
 """
-docstub module docstring
+The same module docstring.
 """
