+++ tests/codemods/codemod_test_cases/mod_merge_class_func/before.py
