+++ tests/codemods/codemod_test_cases/mod_merge_class_func/expected.py
@@ -1,0 +1,216 @@
+"""
+network configuration. See: https://docs.micropython.org/en/v1.18/library/network.html
+
+This module provides network drivers and routing configuration. To use this
+module, a MicroPython variant/build with network capabilities must be installed.
+Network drivers for specific hardware are available within this module and are
+used to configure hardware network interface(s). Network services provided
+by configured interfaces are then available for use via the :mod:`socket`
+module.
+
+For example::
+
+    # connect/ show IP config a specific network interface
+    # see below for examples of specific drivers
+    import network
+    import time
+    nic = network.Driver(...)
+    if not nic.isconnected():
+        nic.connect()
+        print("Waiting for connection...")
+        while not nic.isconnected():
+            time.sleep(1)
+    print(nic.ifconfig())
+
+    # now use socket as usual
+    import socket
+    addr = socket.getaddrinfo('micropython.org', 80)[0][-1]
+    s = socket.socket()
+    s.connect(addr)
+    s.send(b'GET / HTTP/1.1
+Host: micropython.org
+
+')
+    data = s.recv(1000)
+    s.close()
+-----
+Module: 'network' on micropython-v1.18-esp32
+"""
+# MCU: {'ver': 'v1.18', 'port': 'esp32', 'arch': 'xtensawin', 'sysname': 'esp32', 'release': '1.18.0', 'name': 'micropython', 'mpy': 10757, 'version': '1.18.0', 'machine': 'ESP32 module (spiram) with ESP32', 'build': '', 'nodename': 'esp32', 'platform': 'esp32', 'family': 'micropython'}
+# Stubber: 1.5.4
+from typing import Callable, Coroutine, Dict, Generator, IO, Iterator, List, NoReturn, Optional, Tuple, Union, Any
+
+AP_IF = 1  # type: int
+AUTH_MAX = 8  # type: int
+AUTH_OPEN = 0  # type: int
+AUTH_WEP = 1  # type: int
+AUTH_WPA2_ENTERPRISE = 5  # type: int
+AUTH_WPA2_PSK = 3  # type: int
+AUTH_WPA2_WPA3_PSK = 7  # type: int
+AUTH_WPA3_PSK = 6  # type: int
+AUTH_WPA_PSK = 2  # type: int
+AUTH_WPA_WPA2_PSK = 4  # type: int
+ETH_CONNECTED = 3  # type: int
+ETH_DISCONNECTED = 4  # type: int
+ETH_GOT_IP = 5  # type: int
+ETH_INITIALIZED = 0  # type: int
+ETH_STARTED = 1  # type: int
+ETH_STOPPED = 2  # type: int
+
+
+def LAN(*args, **kwargs) -> Any:
+    ...
+
+
+MODE_11B = 1  # type: int
+MODE_11G = 2  # type: int
+MODE_11N = 4  # type: int
+PHY_DP83848 = 3  # type: int
+PHY_IP101 = 1  # type: int
+PHY_LAN8720 = 0  # type: int
+PHY_RTL8201 = 2  # type: int
+
+
+def PPP(*args, **kwargs) -> Any:
+    ...
+
+
+STAT_ASSOC_FAIL = 203  # type: int
+STAT_BEACON_TIMEOUT = 200  # type: int
+STAT_CONNECTING = 1001  # type: int
+STAT_GOT_IP = 1010  # type: int
+STAT_HANDSHAKE_TIMEOUT = 204  # type: int
+STAT_IDLE = 1000  # type: int
+STAT_NO_AP_FOUND = 201  # type: int
+STAT_WRONG_PASSWORD = 202  # type: int
+STA_IF = 0  # type: int
+
+class WLAN(AbstractNIC):
+    """
+    Create a WLAN network interface object. Supported interfaces are
+    ``network.STA_IF`` (station aka client, connects to upstream WiFi access
+    points) and ``network.AP_IF`` (access point, allows other WiFi clients to
+    connect). Availability of the methods below depends on interface type.
+    For example, only STA interface may `WLAN.connect()` to an access point.
+    """
+
+    def __init__(self, interface_id) -> None: ...
+    def active(self, is_active: Optional[Any] = None) -> None:
+        """
+        Activate ("up") or deactivate ("down") network interface, if boolean
+        argument is passed. Otherwise, query current state if no argument is
+        provided. Most other methods require active interface.
+        """
+        ...
+    def connect(self, ssid=None, password=None, *, bssid=None) -> None:
+        """
+        Connect to the specified wireless network, using the specified password.
+        If *bssid* is given then the connection will be restricted to the
+        access-point with that MAC address (the *ssid* must also be specified
+        in this case).
+        """
+        ...
+    def disconnect(self) -> None:
+        """
+        Disconnect from the currently connected wireless network.
+        """
+        ...
+    def scan(self) -> List[Tuple]:
+        """
+        Scan for the available wireless networks.
+        Hidden networks -- where the SSID is not broadcast -- will also be scanned
+        if the WLAN interface allows it.
+
+        Scanning is only possible on STA interface. Returns list of tuples with
+        the information about WiFi access points:
+
+            (ssid, bssid, channel, RSSI, authmode, hidden)
+
+        *bssid* is hardware address of an access point, in binary form, returned as
+        bytes object. You can use `binascii.hexlify()` to convert it to ASCII form.
+
+        There are five values for authmode:
+
+            * 0 -- open
+            * 1 -- WEP
+            * 2 -- WPA-PSK
+            * 3 -- WPA2-PSK
+            * 4 -- WPA/WPA2-PSK
+
+        and two for hidden:
+
+            * 0 -- visible
+            * 1 -- hidden
+        """
+        ...
+    def status(self, param: Optional[Any] = None) -> Any:
+        """
+        Return the current status of the wireless connection.
+
+        When called with no argument the return value describes the network link status.
+        The possible statuses are defined as constants:
+
+            * ``STAT_IDLE`` -- no connection and no activity,
+            * ``STAT_CONNECTING`` -- connecting in progress,
+            * ``STAT_WRONG_PASSWORD`` -- failed due to incorrect password,
+            * ``STAT_NO_AP_FOUND`` -- failed because no access point replied,
+            * ``STAT_CONNECT_FAIL`` -- failed due to other problems,
+            * ``STAT_GOT_IP`` -- connection successful.
+
+        When called with one argument *param* should be a string naming the status
+        parameter to retrieve.  Supported parameters in WiFI STA mode are: ``'rssi'``.
+        """
+        ...
+    def isconnected(self) -> bool:
+        """
+        In case of STA mode, returns ``True`` if connected to a WiFi access
+        point and has a valid IP address.  In AP mode returns ``True`` when a
+        station is connected. Returns ``False`` otherwise.
+        """
+        ...
+    def ifconfig(self, configtuple: Optional[Any] = None) -> Tuple:
+        """
+        Get/set IP-level network interface parameters: IP address, subnet mask,
+        gateway and DNS server. When called with no arguments, this method returns
+        a 4-tuple with the above information. To set the above values, pass a
+        4-tuple with the required information.  For example::
+
+         nic.ifconfig(('192.168.0.4', '255.255.255.0', '192.168.0.1', '8.8.8.8'))
+        """
+        ...
+    def config(self, param) -> Any:
+        """
+        Get or set general network interface parameters. These methods allow to work
+        with additional parameters beyond standard IP configuration (as dealt with by
+        `WLAN.ifconfig()`). These include network-specific and hardware-specific
+        parameters. For setting parameters, keyword argument syntax should be used,
+        multiple parameters can be set at once. For querying, parameters name should
+        be quoted as a string, and only one parameter can be queries at time::
+
+         # Set WiFi access point name (formally known as ESSID) and WiFi channel
+         ap.config(essid='My AP', channel=11)
+         # Query params one by one
+         print(ap.config('essid'))
+         print(ap.config('channel'))
+
+        Following are commonly supported parameters (availability of a specific parameter
+        depends on network technology type, driver, and :term:`MicroPython port`).
+
+        =============  ===========
+        Parameter      Description
+        =============  ===========
+        mac            MAC address (bytes)
+        essid          WiFi access point name (string)
+        channel        WiFi channel (integer)
+        hidden         Whether ESSID is hidden (boolean)
+        authmode       Authentication mode supported (enumeration, see module constants)
+        password       Access password (string)
+        dhcp_hostname  The DHCP hostname to use
+        reconnects     Number of reconnect attempts to make (integer, 0=none, -1=unlimited)
+        =============  ===========
+        """
+        ...
+
+
+def phy_mode(*args, **kwargs) -> Any:
+    ...
